Converted to use OpenGL style matrix

+++ src/main/kotlin/com/zynaps/graphics/Edge.kt
@@ -1,7 +1,7 @@
 package com.zynaps.graphics
 
-import com.zynaps.math.Scalar.ceil
 import java.lang.Math.fma
+import kotlin.math.ceil
 
 internal class Edge {
     var y = 0
@@ -20,10 +20,10 @@ internal class Edge {
     var lOverZStep = 0F
 
     fun compute(a: Vertex, b: Vertex, gradients: Gradients, height: Int): Int {
-        y = ceil(a.y)
+        y = ceil(a.y).toInt()
         if (y < 0) y = 0
 
-        this.height = ceil(b.y)
+        this.height = ceil(b.y).toInt()
         if (this.height > height) this.height = height
 
         this.height -= y
